finish binary tree
